informing user when there is problem with replay
added messageboxes and checking for existing directory/replay/etc.

+++ FileManager/FileManager_frm.cs
@@ -89,8 +89,22 @@ namespace YGOPro_Launcher
                     return;
                 }
                 string replayDir = Path.GetDirectoryName(Application.ExecutablePath) + "/replay/";
+                if (!Directory.Exists(replayDir))
+                {
+                    MessageBox.Show("Replay directory doesn't exist!");
+                    return;
+                }
                 string fileName = replayDir + FileList.SelectedItem + ".yrp";
+                if (!File.Exists(fileName))
+                {
+                    MessageBox.Show("Selected file does not exist!");
+                    return;
+                }
                 string tempFile = replayDir + "000000000000000000000000000000000000000000000.yrp";
+                if (File.Exists(tempFile))
+                {
+                    File.Delete(tempFile);
+                }
                 File.Copy(fileName, tempFile);
                 LauncherHelper.RunGame("-r", (evSender, evArgs) => File.Delete(tempFile));
             }
